refactor: move supervised model into module
