+++ src/model/mod.rs
@@ -1,0 +1,5 @@
+//! Types and implementations for machine learning models.
+
+pub mod supervised;
+
+pub use supervised::SupervisedModel;
